requirements.txt
qdrant-client==1.7.0
sentence-transformers==2.6.1
PyPDF2==3.0.1
nltk==3.8.1
pyttsx3==2.90
requests==2.31.0
openai==1.30.1
pyttsx3==2.90
python-dotenv==1.0.1


qdrant-client
sentence-transformers
PyPDF2
nltk
pyttsx3
requests
